Check for existence of patch.diff instead of deleting it right away

+++ tools/applyformatpatch.sh
@@ -1,6 +1,9 @@
 #!/bin/sh
 
-rm -f patch.diff
+if test -e patch.diff ; then
+  rm -f patch.diff
+  echo patch.diff was removed
+fi
 
 BRANCH=$(git rev-parse --abbrev-ref HEAD)
 RUNID=$(gh run list -b "$BRANCH" -w "Build and Test" --json databaseId -q .[0].databaseId)
